+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.1.5 ⚙️ ongoing work
 
+* Add command to kill specific hap / all active haps
+
 
 ## 0.1.4 📦 current package